==> 1_data_load_0225.ipynb <==
################## preprocessing ###################
#################################################### 
####################################################
import pandas as pd
from os.path import isfile, join
import os
import logging
import argparse
from sklearn.preprocessing import LabelEncoder, MinMaxScaler

date = '20220126'
basedir = '/home/lms/ms/poc'
rawdata_path = join(basedir, 'dataset')
procdata_path = join(basedir, 'deepfm', 'dataset', 'preprocess', 'rating')
procdata_path_r = join(basedir, 'deepfm', 'dataset', 'preprocess', 'rating')

user_data = pd.read_pickle(os.path.join(rawdata_path, 'user.pkl')) # unique id = 2047776
item_data = pd.read_pickle(os.path.join(rawdata_path, 'item.pkl'))
hist_data = pd.read_pickle(os.path.join(rawdata_path, 'history.pkl')) # unique id = 2038827

# 데이터 타입 통일
user_data = user_data.astype({'sa_id':'str'})
item_data = item_data.astype({'category_id':'str', 'category_name':'str', 'album_id':'str'})
hist_data = hist_data.astype({'sa_id':'str', 'album_id':'str', 'category_id':'str'})

item_columns = ['category_id','category_name','pr_info','price','release_date','run_time_ss']
item_data = item_data.dropna(subset=item_columns)

############################### 22.01.24 rating 신규 ##########################
# hist_data['rating_5'] = 5 # 5점 만들기
hist_data['rating_r'] = 2.5 + hist_data['watch_ratio'] * 2.5  # 나름대로 조작

hist_data.head(2)

# 차이 존재
hist_data = hist_data[['sa_id','album_id', 'rating_r']]

hist_data1 = hist_data.merge(item_data[['album_id','category_id']], on='album_id')

train_data = pd.DataFrame(hist_data1.groupby(by=['sa_id','category_id'], sort=False)['rating_r'].max()).reset_index()   # Maximun

train_data['rating_5'] = 5

item_columns = ['pr_info','price','release_date','run_time_ss']
category_data = item_data.groupby(by='category_id', sort=False)[item_columns].max()
category_data = pd.DataFrame(category_data).reset_index()

item_columns = ['category_id','pr_info','price','release_date','run_time_ss']
train_data = train_data.merge(user_data, on='sa_id')

train_data = train_data.merge(category_data[item_columns], on='category_id')

def encoding(dataset, columns):
    # label encoding
    for col_name in columns:
        codes, _ = dataset[col_name].factorize()            
        dataset[col_name + '_CODE'] = codes            

    return dataset

def minmax_scaling(dataset, columns):
    mms = MinMaxScaler(feature_range=(0, 1))
    dataset[columns] = mms.fit_transform(dataset[columns])

    return dataset

train_data = encoding(train_data, ['category_id','pr_info','price'])
train_data = encoding(train_data, ['sa_id'])

scale_columns = ['release_date', 'run_time_ss']
train_data['release_date'] = train_data['release_date'].replace('', '19900101')
train_data[scale_columns] = minmax_scaling(train_data[scale_columns], scale_columns)

# for mapping data
mapping = train_data[['sa_id', 'category_id', 'category_id_CODE', 'sa_id_CODE']]
mapping.to_pickle(os.path.join(procdata_path_r, 'mapping.pkl'))

def save_code(dataset, user_column='sa_id', item_column='category_id'):
    user_map = dataset[[user_column, 
                        user_column+'_CODE', 
                        'seg_1', 
                        'seg_2', 
                        'seg_3', 
                        'seg_4']].drop_duplicates()
    item_map = dataset[[item_column, 
                        item_column+'_CODE', 
                        'price_CODE', 
                        'pr_info_CODE', 
                        'release_date', 
                        'run_time_ss']].drop_duplicates()

    user_map.to_pickle(os.path.join(procdata_path_r, 'user.pkl'))
    item_map.to_pickle(os.path.join(procdata_path_r, 'item.pkl'))
            

def save_history(dataset):
    dataset.to_pickle(os.path.join(procdata_path_r, 'hist.pkl'))

save_code(train_data)

save_columns = ['sa_id_CODE',
                'category_id_CODE',
                'rating_r',
                'rating_5',
                'seg_1',
                'seg_2',
                'seg_3',
                'seg_4',
                'pr_info_CODE',
                'price_CODE',
                'release_date',
                'run_time_ss'
               ]
save_history(train_data[save_columns])
