==> 2_train_0225_r_f.ipynb <==
import os
import tqdm
import numpy as np
import pandas as pd
import tensorflow as tf
from os.path import isfile, join
import time
import math
import logging
from tqdm import tqdm, tqdm_pandas 

from deepctr.models import DeepFM
from deepctr.feature_column import SparseFeat,DenseFeat, get_feature_names, VarLenSparseFeat

# gpu number setting

os.environ['CUDA_DEVICE_ORDER'] = 'PCI_BUS_ID'
os.environ["CUDA_VISIBLE_DEVICES"] = '0'

config = tf.compat.v1.ConfigProto()
config.gpu_options.allow_growth = True
session = tf.compat.v1.Session(config=config)

# tensorflow & keras version check
print('tensorflow version : ' , tf.__version__)
print('keras version : ' , tf.keras.__version__)

# tensorflow gpu available check 
print('GPU available ? : ', tf.test.is_gpu_available(cuda_only=False, min_cuda_compute_capability=None))


basedir = '/home/lms/ms/poc'
rawdata_path = join(basedir, 'dataset')
procdata_path = join(basedir, 'deepfm', 'dataset', 'preprocess', 'rating')

model_path_5_no_f= join(basedir, 'deepfm', 'model', 'model_5_no_f')
model_path_r_no_f= join(basedir, 'deepfm', 'model', 'model_r_no_f')
model_path_5_f= join(basedir, 'deepfm', 'model', 'model_5_f')
model_path_r_f= join(basedir, 'deepfm', 'model', 'model_r_f')

target_5 = 'rating_5'
target_r = 'rating_r'
epochs=20
batch_size=2**15
learner='adam'
num_hidden=256
task='regression'
dropout=0.6
lr=0.001
lr_decay=0.0001
embedding_dim=5
data_filename='hist.pkl'

data = pd.read_pickle(os.path.join(procdata_path, data_filename))
data.head()

sparse_features = ['sa_id_CODE','category_id_CODE','price_CODE','pr_info_CODE'] # nominal variable
dense_features = ['seg_1',
                  'seg_2',
                  'seg_3',
                  'seg_4',                       
                  'release_date',
                  'run_time_ss'] # continuous 

fixlen_feature_columns = [SparseFeat(feat, data[feat].nunique(), embedding_dim=embedding_dim) 
                                  for feat in sparse_features] + [DenseFeat(feat, 1,) for feat in dense_features]

linear_feature_columns = fixlen_feature_columns
dnn_feature_columns = fixlen_feature_columns

feature_names = get_feature_names(linear_feature_columns + dnn_feature_columns)

fixlen_feature_columns

data = data[['sa_id_CODE', 'category_id_CODE', 'rating_r', 'seg_1', 'seg_2', 'seg_3', 'seg_4', 'pr_info_CODE', 'price_CODE', 'release_date', 'run_time_ss']]
data.head()

from sklearn.metrics import mean_squared_error
from sklearn.model_selection import train_test_split

train, test = train_test_split(data, test_size=0.2, random_state=2020)

# negative random sampling
max_code = data['category_id_CODE'].values.tolist()
max_num = max(max_code)

np.random.seed = 2020
rand1 = np.random.randint(max_num, size=len(train)) # 12511
rand2 = np.random.randint(max_num, size=len(train))
rand3 = np.random.randint(max_num, size=len(train))
rand4 = np.random.randint(max_num, size=len(train))
rand5 = np.random.randint(max_num, size=len(train))

train0 = pd.DataFrame(columns=['sa_id_CODE', 'category_id_CODE'])
train0['sa_id_CODE'] = train['sa_id_CODE']
train0

train1 = train0.copy()
train2 = train0.copy()
train3 = train0.copy()
train4 = train0.copy()
train5 = train0.copy()

train1['category_id_CODE'] = rand1
train1['rating_r'] = 0
train2['category_id_CODE'] = rand2
train2['rating_r'] = 0
train3['category_id_CODE'] = rand3
train3['rating_r'] = 0
train4['category_id_CODE'] = rand4
train4['rating_r'] = 0
train5['category_id_CODE'] = rand5
train5['rating_r'] = 0

train = train[['sa_id_CODE', 'category_id_CODE', 'rating_r']]

train = pd.concat([train, train1, train2, train3, train4, train5]) # 77011698

# user feature merge
user_data = pd.read_pickle(os.path.join(procdata_path, 'user.pkl'))
user_data = user_data[['sa_id_CODE', 'seg_1', 'seg_2', 'seg_3', 'seg_4']]
# item feature merge
item_data = pd.read_pickle(os.path.join(procdata_path, 'item.pkl'))
item_data = item_data[['category_id_CODE', 'pr_info_CODE', 'price_CODE', 'release_date', 'run_time_ss']]

train = train.merge(user_data, on='sa_id_CODE')

train = train.merge(item_data, on='category_id_CODE')

del [[train1, train2, train3, train4, train5, train0, data]]
#del train2
#del train3
#del train4
#del train5
#del train0
#del data
#del user_data
#del item_data

# 셔플
train = train.sample(frac=1).reset_index(drop=True)

# save
train.to_pickle(os.path.join(procdata_path, 'train_r.pkl'))

# save
test.to_pickle(os.path.join(procdata_path, 'test_r.pkl'))

del [[train, test]]

train = pd.read_pickle(os.path.join(procdata_path, 'train_r.pkl'))
test = pd.read_pickle(os.path.join(procdata_path, 'test_r.pkl'))

train_model_input = {name:train[name].values for name in feature_names}
test_model_input = {name:test[name].values for name in feature_names}

train_model_input

model = DeepFM(linear_feature_columns,
                            dnn_feature_columns,
                            dnn_hidden_units=(num_hidden, num_hidden),
                            dnn_dropout=dropout,
                            task=task)
        
if learner.lower() == "adagrad": 
    opt = tf.keras.optimizers.Adagrad(lr=lr,  decay=lr_decay)
elif learner.lower() == "rmsprop":
    opt = tf.keras.optimizers.RMSprop(lr=lr,  decay=lr_decay)
elif learner.lower() == "adam":
    opt = tf.keras.optimizers.Adam(lr=lr, decay=lr_decay)
else:
    opt = tf.keras.optimizers.SGD(lr=lr,  decay=lr_decay)

model.compile(opt, 'mse', metrics=['mse',tf.keras.metrics.RootMeanSquaredError()])

train

epochs = 7
best_loss = epochs
        
for epoch in range(epochs):

    # Training
    hist = model.fit(train_model_input, #input
                          train[target_r].values, # labels
                          validation_data=(test_model_input, test[target_r].values),
                          batch_size=batch_size,
                          initial_epoch=epoch, epochs=epoch + 1, 
                          verbose=1, 
                          shuffle=True)

    train_loss = hist.history['loss'][0]
    train_rmse = hist.history['root_mean_squared_error'][0]
    val_loss = hist.history['val_loss'][0]
    val_rmse = hist.history['val_root_mean_squared_error'][0]


    if val_loss < best_loss :
        best_loss = val_loss
        model.save_weights(os.path.join(model_path_r_f, 'best.h5'))

model.save_weights(os.path.join(model_path_r_f, 'best.h5'))

model.load_weights(os.path.join(model_path_r_f, 'best.h5'))
tf.keras.models.save_model(model, os.path.join(model_path_r_f, 'bestmodel'))
