==> 4_score_negative.ipynb <==
# scoreing 파일

import os
import math
from tqdm import tqdm_notebook as tqdm
import time
import numpy as np
import pandas as pd
import logging
import argparse

import warnings
import multiprocessing
warnings.filterwarnings("ignore")

from sklearn.metrics import mean_squared_error
from sklearn.model_selection import train_test_split

from deepctr.models import DeepFM
from deepctr.feature_column import SparseFeat,DenseFeat, get_feature_names, VarLenSparseFeat

import os
from os.path import isfile, join
basedir = '/home/lms/ms/poc'
rawdata_path = join(basedir, 'dataset')
procdata_path = join(basedir, 'deepfm', 'dataset', 'preprocess', 'rating')
result_path_5_no_f = join(basedir, 'deepfm', 'result', 'result_5_no_f') # path 설정 필요
result_path_r_no_f = join(basedir, 'deepfm', 'result', 'result_r_no_f') # path 설정 필요
result_path_5_f = join(basedir, 'deepfm', 'result', 'result_5_f') # path 설정 필요
result_path_r_f = join(basedir, 'deepfm', 'result', 'result_r_f') # path 설정 필요

model_path_5_no_f= join(basedir, 'deepfm', 'model', 'model_5_no_f')
model_path_r_no_f= join(basedir, 'deepfm', 'model', 'model_r_no_f')
model_path_5_f= join(basedir, 'deepfm', 'model', 'model_5_f')
model_path_r_f= join(basedir, 'deepfm', 'model', 'model_r_f')

batch_size = 200
topk = 25

# 추천리스트 load
df0 = pd.read_csv(os.path.join(result_path_5_f, 'reclist_0.csv'), sep='\001', header=None)
df0.head()

#load test data
from sklearn.metrics import mean_squared_error
from sklearn.model_selection import train_test_split
data = pd.read_pickle(os.path.join(procdata_path, 'hist.pkl'))
train, test = train_test_split(data, test_size=0.2, random_state=2020)
test.head()

test_set = test.sort_values(by=['sa_id_CODE', 'category_id_CODE'], ascending=[True, True], axis=0)
test_set.head()

# test set에 있는 유저 추출
user_set = test_set['sa_id_CODE'].unique()
user_set = list(user_set)
user_set = [ int(x) for x in user_set]

# test set에 있는 유저에 대한 추천리스트 추출
pred = df0[df0[1].isin(user_set)]
pred.head()

# 참고.....
df1 = pd.read_csv(os.path.join(model_pred_path, date, 'reclist_1.csv'), sep='\001', header=None)
df2 = pd.read_csv(os.path.join(model_pred_path, date, 'reclist_2.csv'), sep='\001', header=None)
# test set2 에 있는 유저에 대한 추천리스트 추출
pred_set2 = df1[df1[1].isin(user_set)]
pred_set3 = df2[df2[1].isin(user_set)]
pred1_set2 = pred_set2.groupby(1)[2].apply(list)
pred1_set3 = pred_set3.groupby(1)[2].apply(list)
predict_set2 = []
for i in pred1_set2:
    predict_set2.append(i)
predict_set3 = []
for i in pred1_set3:
    predict_set3.append(i)

pred1 = pred.groupby(1)[2].apply(list)

# 추천리스트를 리스트로
predict = []
for i in pred1:
    predict.append(i)

# test set을 리스트로
la = test_set.groupby('sa_id_CODE')['category_id_CODE'].apply(list)
label = []
for i in la:
    label.append(i)

# 전체 testset 중 일부만 (현재는 약 절반) 대상
target = label[:len(predict)]
len(target) #####

### 지표 시작!!!

def calc_precision(predicted, actual):
        prec = [value for value in predicted if value in actual]
        prec = np.round(float(len(prec)) / float(len(predicted)), 4)
        return prec

pred_at = np.mean(list(map(calc_precision, predict, target)))

pred_at ################## precision

def calc_recall(predicted, actual):
    reca = [value for value in predicted if value in actual]
    reca = np.round(float(len(reca)) / float(len(actual)), 4)
    return reca

recall_at = np.mean(list(map(calc_recall, predict, target)))

recall_at  ###################### recall

def prediction_coverage(predicted, tot):
    predicted_flattened = [p for sublist in predicted for p in sublist]
    unique_predictions = len(set(predicted_flattened))
    prediction_coverage = round(unique_predictions/tot*100,2)
    return prediction_coverage

tot = data['category_id_CODE'].nunique()

cov_at = prediction_coverage(predict,tot)

print(cov_at, "%")  ############### coverage

def mrr(prediction, label):
    rr = 0
    for i in range(len(prediction)):
        if any(np.isin(prediction[i], label[i]))==False:
            continue
        else:
            rank = np.min(np.where(np.isin(prediction[i], label[i])))+1
        rr = rr + 1/rank
    return rr/len(prediction)

mrr_at = mrr(predict, target)
mrr_at ########################### mrr

### negative sampling 테스트 (p@25 / r@25 / c@25 / m@25)
#### rating_5_no_f : 0.032 / 0.29 / 0.09 / 0.15
#### rating_r_no_f : 0.030 / 0.27 / 0.13 / 0.14 
#### rating_5_f : 0.034 / 0.31 / 0.09 / 0.15
#### rating_r_f : 0.031 / 0.29 / 0.09 / 0.14

# 기존 상용 지표
### 평점 : (상용 모델) 시청시간 + 최신성 => 0.0035 / 0.027 / 0.46% / 0.012
### 평점 : 0.5 + 시청시간 + 최신성 => 0.005 / 0.04 / 0.6% / 0.015 (1/24)
### 평점 : 5 (epoch 5) => 0.0001 / 0.0004 / 1.11% / 0.0001 (1/25)
### 평점 : 5 (epoch 10) => 안좋음... (1/25_1)
### 평점 : 0.5 + 시청 시간 + 최신성 + 네버티브 샘플링(0.1, 5개) => 0.0078 / 0.068 / 0.71% / 0.0198

### 2번째평가 함수들

def precision_at(predictions, labels, k=10, assume_unique=True):
    # validate K
    # _require_positive_k(k)

    def _inner_pk(pred, lab):
        # need to compute the count of the number of values in the predictions
        # that are present in the labels. We'll use numpy in1d for this (set
        # intersection in O(1))
        if lab.shape[0] > 0:
            n = min(pred.shape[0], k)
            cnt = np.in1d(pred[:n], lab, assume_unique=assume_unique).sum()
            return float(cnt) / k
        else:
            return _warn_for_empty_labels()

    return _mean_ranking_metric(predictions, labels, _inner_pk)

def _mean_ranking_metric(predictions, labels, metric):
    """Helper function for precision_at_k and mean_average_precision"""
    # do not zip, as this will require an extra pass of O(N). Just assert
    # equal length and index (compute in ONE pass of O(N)).
    # if len(predictions) != len(labels):
    #     raise ValueError("dim mismatch in predictions and labels!")
    # return np.mean([
    #     metric(np.asarray(predictions[i]), np.asarray(labels[i]))
    #     for i in xrange(len(predictions))
    # ])
    
    # Actually probably want lazy evaluation in case preds is a 
    # generator, since preds can be very dense and could blow up 
    # memory... but how to assert lengths equal? FIXME
    return np.mean([
        metric(np.asarray(prd), np.asarray(labels[i]))
        for i, prd in enumerate(predictions)  # lazy eval if generator
    ])

def mean_average_precision(predictions, labels, assume_unique=True):
    def _inner_map(pred, lab):
        if lab.shape[0]:
            # compute the number of elements within the predictions that are
            # present in the actual labels, and get the cumulative sum weighted
            # by the index of the ranking
            n = pred.shape[0]

            # Scala code from Spark source:
            # var i = 0
            # var cnt = 0
            # var precSum = 0.0
            # val n = pred.length
            # while (i < n) {
            #     if (labSet.contains(pred(i))) {
            #         cnt += 1
            #         precSum += cnt.toDouble / (i + 1)
            #     }
            #     i += 1
            # }
            # precSum / labSet.size

            arange = np.arange(n, dtype=np.float32) + 1.  # this is the denom
            present = np.in1d(pred[:n], lab, assume_unique=assume_unique)
            prec_sum = np.ones(present.sum()).cumsum()
            denom = arange[present]
            return (prec_sum / denom).sum() / lab.shape[0]

        else:
            return _warn_for_empty_labels()

    return _mean_ranking_metric(predictions, labels, _inner_map)


def mrr_score(y_true, y_score):
    order = np.argsort(y_score)[::-1]
    y_true = np.take(y_true, order)
    rr_score = y_true / (np.arange(len(y_true)) + 1)
    return np.sum(rr_score) / np.sum(y_true)

pp = [[1,2,3,4,5], [3,4,5,6,7], [1,2,3,4,5]]
aa = [[6], [2,3,7], [4,1]]

rr=0
for i in range(len(pp)):
    rank = np.min(np.where(np.isin(pp[i], aa[i])))+1
    rr = rr + 1/rank

def mrr(prediction, label):
    rr = 0
    for i in range(len(prediction)):
        print(i)
        if any(np.isin(prediction[i], label[i]))==False:
            continue
        else:
            rank = np.min(np.where(np.isin(prediction[i], label[i])))+1
        rr = rr + 1/rank
        print(rank)
    return rr/len(prediction)

# MAP
map_k = mean_average_precision(predict, target)
map_k

### 방법2

## 추론 테스트

df0 = pd.read_csv(os.path.join(model_pred_path, date, 'reclist_0.csv'), sep='\001', header=None)
df0.head()

df1 = pd.read_csv(os.path.join(model_pred_path, date, 'reclist_1.csv'), sep='\001', header=None)
df1.head()

df2 = pd.read_csv(os.path.join(model_pred_path, date, 'reclist_2.csv'), sep='\001', header=None)
df2.head()

df0[df0[5]==1].head(30)

df0[df0[1]==63]

926380 / 1139744 / 370071 / 775896 / 825254 / 1246938

mapping = pd.read_pickle(os.path.join(procdata_path, 'mapping.pkl'))
mapping

mapping[mapping['sa_id']=='500009388201']

mapping[mapping['sa_id_CODE']==63]

items = pd.read_pickle(os.path.join(procdata_path, 'item.pkl'))
item_columns = ['category_id','gubun','rep_album_id','category_name']
category_data = pd.read_pickle(os.path.join(rawdata_path, 'item.pkl'))
category_data = category_data.astype({'category_id':'str', 'category_name':'str', 'album_id':'str'})
category_data = category_data.drop_duplicates(subset=item_columns)
category_data = items.merge(category_data, on='category_id')

map_category = category_data[['category_name', 'category_id_CODE', 'dpeth0_nm']]
map_category

mapping1 = pd.merge(mapping, map_category, how="left", on="category_id_CODE")
mapping1

mapping1[mapping1['sa_id_CODE']==63]

################# 평점 보기

data = pd.read_pickle(os.path.join(procdata_path, 'hist.pkl'))
data.head()

data[data['sa_id_CODE']==370071]
